requirements.txt
greenlet==2.0.2
PyMySQL==1.1.0
python-decouple==3.8
SQLAlchemy==2.0.20
typing_extensions==4.7.1
